refactor(library/MediaType): un-nests directory from URI library
- URIs are just one of the _many_ common use-cases for media identifiers
- ergo a standalone directory is warranted
