+++ src/library/MediaType/StructuredSyntaxNameSuffix/Any.ts
