+++ src/library/MediaType/StructuredSyntaxNameSuffix/all.ts
