+++ src/library/MediaType/StructuredSyntaxNameSuffix/index.ts
