+++ src/library/MediaType/TopLevel/Discrete/Any.ts
